Finalizados cenario 1 e 2

+++ classes/Avaliacao.h
@@ -8,6 +8,6 @@
 struct Avaliacao
 {
       int id;
-      int nota;
+      float nota;
       string nome;
 };
